updates for deleting a folder and setting permissions when adding a subfolder

+++ trunk/ir_service/src/edu/ur/ir/groupspace/service/DefaultGroupWorkspaceFileSystemService.java
@@ -494,7 +494,10 @@ public class DefaultGroupWorkspaceFileSystemService implements GroupWorkspaceFil
 		for( GroupWorkspaceFolder childFolder : folders)
 		{
 			IrAcl acl = securityService.getAcl(childFolder);
-			securityService.deleteAcl(acl);
+			if( acl != null )
+			{
+			    securityService.deleteAcl(acl);
+			}
 		}
 		
 		if( folder.getIsRoot())
